update read values 32 -> 64 bits

+++ stplot-gui/modules/ReadManager/readloop.h
@@ -27,7 +27,7 @@ public:
 public slots:
     void readLoop();
     void stopLoop();
-    void requestWriteData(uint32_t data, varloc_location_t location);
+    void requestWriteData(uint64_t data, varloc_location_t location);
 
 private:
     void calcDecodList(QVector<VarChannel *> *channels);
@@ -50,7 +50,7 @@ private:
     bool stopSignal;
     QVector<ReadDeviceObject::ReadAddres> readSequence;
     QVector<QPair<uint32_t,QVector<uint8_t>>> saveSequence;
-    QVector<QPair<uint32_t,varloc_location_t>> requestedWriteData;
+    QVector<QPair<uint64_t,varloc_location_t>> requestedWriteData;
 
     //map decod chanales bitmap and locations
     QVector<varloc_location_t> decodeList;
